Sort api directory
