--- src/api/default.ts
@@ -1,10 +1,0 @@
-export const defaultFetchConfig: Partial<RequestInit> = {
-  mode: 'cors',
-  cache: 'no-cache',
-  credentials: 'same-origin',
-  headers: {
-    'Content-Type': 'application/json'
-  },
-  redirect: 'follow',
-  referrerPolicy: 'no-referrer'
-}
