+++ src/components/editor/markdown-renderer/replace-components/image/image-frame.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
-import { getProxiedUrl } from '../../../../../api/imageProxy'
+import { getProxiedUrl } from '../../../../../api/media'
 import { ApplicationState } from '../../../../../redux'
 
 export const ImageFrame: React.FC<React.ImgHTMLAttributes<HTMLImageElement>> = ({ alt, src, ...props }) => {
