+++ src/components/landing/pages/history/history.tsx
@@ -3,7 +3,7 @@ import { Row } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { deleteHistory, deleteHistoryEntry, getHistory, setHistory, updateHistoryEntry } from '../../../../api/history'
-import { deleteNote } from '../../../../api/note'
+import { deleteNote } from '../../../../api/notes'
 import { ApplicationState } from '../../../../redux'
 import {
   collectEntries,
