+++ src/components/landing/pages/login/auth/via-internal.tsx
@@ -5,7 +5,7 @@ import { useSelector } from 'react-redux'
 import { Link } from 'react-router-dom'
 import { doInternalLogin } from '../../../../../api/auth'
 import { ApplicationState } from '../../../../../redux'
-import { getAndSetUser } from '../../../../../utils/apiUtils'
+import { getAndSetUser } from '../../../../../api/utils'
 import { ShowIf } from '../../../../common/show-if/show-if'
 
 export const ViaInternal: React.FC = () => {
