+++ src/components/landing/pages/register/register.tsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux'
 import { Redirect } from 'react-router'
 import { doInternalRegister } from '../../../../api/auth'
 import { ApplicationState } from '../../../../redux'
-import { getAndSetUser } from '../../../../utils/apiUtils'
+import { getAndSetUser } from '../../../../api/utils'
 import { Row, Col, Card, Form, Button, Alert } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
 import { TranslatedExternalLink } from '../../../common/links/translated-external-link'
